+++ Frontend/src/app/searchResults/modules/feed/components/feed/feed.component.ts
