+++ Frontend/src/app/searchResults/modules/feed/services/feed.service.ts
