+++ Frontend/src/app/searchResults/modules/feed/store/actionTypes.ts
