+++ Frontend/src/app/searchResults/modules/feed/store/actions/getFeed.action.ts
