+++ Frontend/src/app/searchResults/modules/feed/store/effects/getFeed.effect.ts
