+++ Frontend/src/app/searchResults/modules/feed/store/reducer.ts
