+++ Frontend/src/app/searchResults/modules/feed/store/selectors.ts
