+++ Frontend/src/app/searchResults/modules/feed/types/feedState.interface.ts
