+++ Frontend/src/app/searchResults/modules/globalFeed/components/globalFeed/globalFeed.component.html
@@ -3,7 +3,6 @@
 
 
 <div>
-  adobe app
   <div>
     <mc-feed [apiUrl]="apiUrl"></mc-feed>
   </div>
