+++ Frontend/src/app/searchResults/modules/globalFeed/components/globalFeed/globalFeed.component.ts
