+++ Frontend/src/app/searchResults/modules/globalFeed/globalFeed.module.ts
@@ -2,7 +2,7 @@ import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {GlobalFeedComponent} from "./components/globalFeed/globalFeed.component";
 import {RouterModule, Routes} from "@angular/router";
-import {FeedModule} from "../shared/modules/feed/feed.module";
+import {FeedModule} from "../feed/feed.module";
 
 const routes: Routes = [
     {
